Solution

+++ app/main.py
@@ -1,14 +1,24 @@
 import os
+from types import TracebackType
+from typing import Optional, Type
 
 
 class CleanUpFile:
-    def __init__(self, filename):
-        self.filename = filename
+    def __init__(self, filename: str) -> None:
+        self.filename: str = filename
 
-    def __enter__(self):
+    def __enter__(self) -> "CleanUpFile":
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        os.remove(self.filename)
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> bool:
+        if os.path.exists(self.filename):
+            os.remove(self.filename)
+        return False
+
